% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{annotation}
\alias{annotation}
\alias{annotation}
\title{Annotation of example data}
\format{A data frame with 30 rows and 4 variables}
\usage{
annotation

annotation
}
\description{
A dataset containing the annotation of example data
The variables are as follows:

Result after formating
The variables are as follows:
}
\details{
\itemize{
  \item Run
  \item Channel
  \item Group
  \item BiologicalMixture
}

\itemize{
  \item Protein
  \item PSM
  \item log2Intensity
  \item Run
  \item Channel
  \item Subject
}
}
\keyword{datasets}
