% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designSampleSizeTMT.R
\name{.calculatePower}
\alias{.calculatePower}
\title{Power calculation}
\usage{
.calculatePower(
  desiredFC,
  FDR,
  delta,
  median_sigma_error,
  median_sigma_subject,
  median_sigma_run,
  numSample
)
}
\arguments{
\item{desiredFC}{the range of a desired fold change which includes the lower
and upper values of the desired fold change.}

\item{FDR}{a pre-specified false discovery ratio (FDR) to control the overall
false positive rate. Default is 0.05}

\item{delta}{difference between means (?)}

\item{median_sigma_error}{median of error standard deviation}

\item{median_sigma_subject}{median standard deviation per subject}

\item{numSample}{minimal number of biological replicates per condition.
TRUE represents you require to calculate the sample size for this category,
else you should input the exact number of biological replicates.}
}
\description{
Power calculation
}
\keyword{internal}
