% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_group_comparison.R
\name{.checkContrastMatrix}
\alias{.checkContrastMatrix}
\title{check whether pairwise comparison. If pairwise, generate a contrast matrix.}
\usage{
.checkContrastMatrix(contrast_matrix)
}
\value{
a contrast matrix
}
\description{
check whether pairwise comparison. If pairwise, generate a contrast matrix.
}
\keyword{internal}
