% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_normalize.R
\name{.countRunsWithNorm}
\alias{.countRunsWithNorm}
\title{Utility function: count runs with "Norm" channel}
\usage{
.countRunsWithNorm(run, condition)
}
\arguments{
\item{run}{vector of run labels}

\item{condition}{vector of condition labels}
}
\value{
integer
}
\description{
Utility function: count runs with "Norm" channel
}
\keyword{internal}
