% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_docs.R
\name{.documentFunction}
\alias{.documentFunction}
\title{A dummy function to store shared documentation items.}
\usage{
.documentFunction(
  fewMeasurements,
  useUniquePeptide,
  summaryforMultipleRows,
  removeProtein_with1Feature,
  removeProtein_with1Protein,
  removeOxidationMpeptides,
  removeMpeptides
)
}
\arguments{
\item{use_log_file}{logical. If TRUE, information about data processing
will be saved to a file.}

\item{append}{logical. If TRUE, information about data processing will be added
to an existing log file.}

\item{verbose}{logical. If TRUE, information about data processing wil be printed
to the console.}

\item{log_file_path}{character. Path to a file to which information about 
data processing will be saved. 
If not provided, such a file will be created automatically.
If `append = TRUE`, has to be a valid path to a file.}
}
\value{
NULL.
}
\description{
A dummy function to store shared documentation items.
}
\keyword{internal}
