% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designSampleSizeTMT.R
\name{.getMedianSigmaRun}
\alias{.getMedianSigmaRun}
\title{Get median per subject or group by subject}
\usage{
.getMedianSigmaRun(var_component)
}
\arguments{
\item{var_component}{data.frame, output of .getVarComponent}
}
\description{
Get median per subject or group by subject
}
\keyword{internal}
