% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designSampleSizeTMT.R
\name{.getMedianSigmaSubject}
\alias{.getMedianSigmaSubject}
\title{Get median per run or run by mix}
\usage{
.getMedianSigmaSubject(var_component)
}
\arguments{
\item{var_component}{data.frame, output of .getVarComponent}
}
\description{
Get median per run or run by mix
}
\keyword{internal}
