% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_normalize.R
\name{.getNormalizationAbundance}
\alias{.getNormalizationAbundance}
\title{Utility function: get mean abundance for "Norm" channels}
\usage{
.getNormalizationAbundance(abundance, condition)
}
\arguments{
\item{abundance}{vector of abundances}

\item{condition}{vector of condition labels}
}
\value{
numeric
}
\description{
Utility function: get mean abundance for "Norm" channels
}
\keyword{internal}
