% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designSampleSizeTMT.R
\name{.getNumSample}
\alias{.getNumSample}
\title{Get sample size}
\usage{
.getNumSample(
  desiredFC,
  power,
  alpha,
  delta,
  median_sigma_error,
  median_sigma_subject,
  median_sigma_run
)
}
\arguments{
\item{desiredFC}{the range of a desired fold change which includes the lower
and upper values of the desired fold change.}

\item{power}{a pre-specified statistical power which defined as the probability
of detecting a true fold change. TRUE represent you require to calculate the power
for this category, else you should input the average of power you expect. Default is 0.9}

\item{alpha}{significance level}

\item{delta}{difference between means (?)}

\item{median_sigma_error}{median of error standard deviation}

\item{median_sigma_subject}{median standard deviation per subject}
}
\description{
Get sample size
}
\keyword{internal}
