% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_normalize.R
\name{.getRunsMedian}
\alias{.getRunsMedian}
\title{Utility function: get median from unique values per run}
\usage{
.getRunsMedian(input)
}
\arguments{
\item{input}{data.table / list}
}
\value{
numeric
}
\description{
Utility function: get median from unique values per run
}
\keyword{internal}
