% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designSampleSizeTMT.R
\name{.getVarComponentTMT}
\alias{.getVarComponentTMT}
\title{Get variances from models fitted by the groupComparison function}
\usage{
.getVarComponentTMT(fitted_models)
}
\arguments{
\item{fitted_models}{FittedModels element of groupComparison output}
}
\description{
Get variances from models fitted by the groupComparison function
}
\keyword{internal}
