% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_group_comparison.R
\name{.handleSingleContrastTMT}
\alias{.handleSingleContrastTMT}
\title{perform statistical inference for single protein and single contrast}
\usage{
.handleSingleContrastTMT(
  contrast,
  fit,
  single_protein,
  coefs,
  protein,
  groups,
  s2_posterior,
  rho,
  vss,
  df_prior,
  s2_df
)
}
\description{
perform statistical inference for single protein and single contrast
}
\keyword{internal}
