% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_methods.R
\name{.logSum}
\alias{.logSum}
\title{Utility function: compute log of sum of 2^x}
\usage{
.logSum(x)
}
\arguments{
\item{x}{numeric}
}
\value{
numeric
}
\description{
Utility function: compute log of sum of 2^x
}
\keyword{internal}
