% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_group_comparison.R
\name{.makeContrastSingleTMT}
\alias{.makeContrastSingleTMT}
\title{Make a contrast}
\usage{
.makeContrastSingleTMT(fit, contrast, single_protein, coefs)
}
\value{
a contrast vector
}
\description{
Make a contrast
}
\keyword{internal}
