% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_methods.R
\name{.makeFactorColumnsTMT}
\alias{.makeFactorColumnsTMT}
\title{Converts required columns to factor in summarization output}
\usage{
.makeFactorColumnsTMT(input)
}
\arguments{
\item{input}{data.table}
}
\value{
a data table with factored columns
}
\description{
Converts required columns to factor in summarization output
}
\keyword{internal}
