% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_normalize.R
\name{.normalizeProteins}
\alias{.normalizeProteins}
\title{Normalization between MS runs (after protein summarization)}
\usage{
.normalizeProteins(input, normalize)
}
\arguments{
\item{input}{data.table}

\item{normalize}{logical, if TRUE, data will be normalized}
}
\value{
data.table
}
\description{
Normalization between MS runs (after protein summarization)
}
\keyword{internal}
