% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_methods.R
\name{.summarizeSimpleStat}
\alias{.summarizeSimpleStat}
\title{Summarize TMT data with a simple aggregate of log-intensities}
\usage{
.summarizeSimpleStat(input, annotation, stat_aggregate)
}
\arguments{
\item{input}{data.table}

\item{annotation}{data.table with run and channel annotation}

\item{stat_aggregate}{function that will be used to compute protein-level
summary}
}
\value{
data.table
}
\description{
Summarize TMT data with a simple aggregate of log-intensities
}
\keyword{internal}
