% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_methods.R
\name{.summarizeTMP}
\alias{.summarizeTMP}
\title{Summarize TMT data with median polish}
\usage{
.summarizeTMP(input, annotation)
}
\arguments{
\item{input}{data.table}

\item{annotation}{data.table with run and channel annotation}
}
\value{
data.table with summaried protein intensities
}
\description{
Summarize TMT data with median polish
}
\keyword{internal}
