% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSummarization.R
\name{getProcessedTMT}
\alias{getProcessedTMT}
\title{Get processed feature-level data}
\usage{
getProcessedTMT(summarized, input)
}
\arguments{
\item{summarized}{output of the MSstatsSummarizeTMT function}

\item{input}{output of MSstatsNormalizeTMT function}
}
\value{
data.table
}
\description{
Get processed feature-level data
}
\keyword{internal}
