% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSummarization.R
\name{getSummarizedTMT}
\alias{getSummarizedTMT}
\title{Get protein-level data from MSstatsSummarizeTMT output}
\usage{
getSummarizedTMT(summarized)
}
\arguments{
\item{summarized}{output of the MSstatsSummarizeTMT function}
}
\value{
data.table
}
\description{
Get protein-level data from MSstatsSummarizeTMT output
}
\keyword{internal}
