% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{groupComparision.results}
\alias{groupComparision.results}
\title{Example of output from groupComparision function}
\format{A data frame with 60 rows and 7 variables.}
\usage{
groupComparision.results
}
\description{
It is calculated from the result of protein.summarizaiton function.
It is the output of groupComparision function
It should includes the required columns as below.
The variables are as follows:
}
\details{
\itemize{
  \item Protein : Protein ID
  \item Label: Label of the pairwise comparision or contrast 
  \item log2FC: Log2 fold change
  \item SE: Standard error of the comparsion of contrast results
  \item DF: Degree of freedom
  \item pvalue: Value of p statistic of the test
}
}
\examples{
head(groupComparision.results)

}
\keyword{datasets}
