% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparsion.TMT.R
\name{groupComparison.TMT}
\alias{groupComparison.TMT}
\title{Summarize PSM level data to protein level}
\usage{
groupComparison.TMT(data, annotation, model)
}
\arguments{
\item{data}{data: protein level data, which has columns Protein, Group, Subject, Run, Channel, IonIntensity}

\item{model}{Possible options: "proposed", "lima", "t"}
}
\value{
The sum of \code{x} and \code{y}.//TODO
}
\description{
Summarize PSM level data to protein level
}
\examples{
//TODO
}
