% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MSstatsTMT.R
\docType{data}
\name{input.pd}
\alias{input.pd}
\title{Example dataset in MSstatsTMT format.}
\format{
A data frame with 20110 rows and 11 variables.
}
\usage{
input.pd
}
\description{
It should include the required columns as below.
}
\details{
\itemize{
  \item ProteinName : Protein ID
  \item PeptideSequence : peptide sequence
  \item Charge : peptide charge
  \item PSM : peptide ion and spectra match
  \item Channel : Labeling information (126, ... 131)
  \item Condition : Condition (ex. Healthy, Cancer, Time0)
  \item BioReplicate : Unique ID for biological subject.
  \item Run : MS run ID
  \item Mixture : Unique ID for TMT mixture.
  \item TechRepMixture : Unique ID for technical replicate of one TMT mixture.
  \item Intensity: Protein Abundance
}
}
\examples{
head(input.pd)

}
\keyword{datasets}
