% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MQtoMSstatsTMTFormat.R
\name{MQtoMSstatsTMTFormat}
\alias{MQtoMSstatsTMTFormat}
\title{Generate MSstatsTMT required input format for MaxQuant output}
\usage{
MQtoMSstatsTMTFormat(evidence, proteinGroups, annotation,
  fraction = FALSE, remove.Only.identified.by.site = FALSE,
  which.proteinid = "Proteins", useUniquePeptide = TRUE,
  summaryforMultipleRows = sum,
  removePSM_withMissingValue_withinRun = TRUE,
  removeProtein_with1Feature = FALSE)
}
\arguments{
\item{evidence}{name of 'evidence.txt' data, which includes feature-level data.}

\item{proteinGroups}{name of 'proteinGroups.txt' data.}

\item{annotation}{data frame which contains column Run, Channel, Condition, BioReplicate, Mixture.}

\item{fraction}{indicates whether the data has fractions. If there are fractions, then overlapped peptide ions will be removed and then fractions are combined for each mixture.}

\item{remove.Only.identified.by.site}{TRUE will remove proteins with '+' in 'Only.identified.by.site' column from proteinGroups.txt, which was identified only by a modification site. FALSE is the default.}

\item{which.proteinid}{Use 'Proteins'(default) column for protein name. 'Leading.proteins' or 'Leading.razor.proteins' can be used instead. However, those can potentially have the shared peptides.}

\item{useUniquePeptide}{TRUE(default) removes peptides that are assigned for more than one proteins. We assume to use unique peptide for each protein.}

\item{summaryforMultipleRows}{max(default) or sum - when there are multiple measurements for certain feature and certain run, use highest or sum of multiple intensities.}

\item{removePSM_withMissingValue_withinRun}{TRUE(default) will remove PSM with any missing value within each Run.}

\item{removeProtein_with1Feature}{TRUE(default) will remove the proteins which have only 1 peptide and charge.}
}
\value{
input for protein.summarization function
}
\description{
Convert MaxQuant output into the required input format for MSstatsTMT.
}
\examples{
\dontrun{
head(evidence)
head(proteinGroups)
head(annotation)
required.input <- MQtoMSstatsTMTFormat(evidence, proteinGroups, annotation)
head(required.input)
}
}
