% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonTMT.R
\name{MSstatsComparisonModelSingleTMT}
\alias{MSstatsComparisonModelSingleTMT}
\title{Fit a linear model for group comparison for a single protein}
\usage{
MSstatsComparisonModelSingleTMT(single_protein, protein_name)
}
\arguments{
\item{single_protein}{protein-level data for a single protein (single element
of list created by the MSstatsPrepareForGroupComparisonTMT function)}

\item{protein_name}{name of a protein from the single_protein data.table}
}
\value{
list
}
\description{
Fit a linear model for group comparison for a single protein
}
\keyword{internal}
