% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonTMT.R
\name{MSstatsFitComparisonModelsTMT}
\alias{MSstatsFitComparisonModelsTMT}
\title{Fit linear models for group comparison}
\usage{
MSstatsFitComparisonModelsTMT(input)
}
\arguments{
\item{input}{output of the MSstatsPrepareForGroupComparisonTMT function}
}
\value{
list
}
\description{
Fit linear models for group comparison
}
\keyword{internal}
