% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonTMT.R
\name{MSstatsGroupComparisonOutputTMT}
\alias{MSstatsGroupComparisonOutputTMT}
\title{Combine testing results for individual proteins}
\usage{
MSstatsGroupComparisonOutputTMT(testing_results, adj_method)
}
\arguments{
\item{testing_results}{output of the MSstatsGroupComparisonTMT function}

\item{adj_method}{method that will be used to adjust p-values for multiple comparisons}
}
\value{
data.table
}
\description{
Combine testing results for individual proteins
}
\keyword{internal}
