% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonTMT.R
\name{MSstatsGroupComparisonTMT}
\alias{MSstatsGroupComparisonTMT}
\title{Group comparison for TMT data}
\usage{
MSstatsGroupComparisonTMT(fitted_models, contrast_matrix)
}
\arguments{
\item{fitted_models}{output of the MSstatsModerateTTest function}

\item{contrast_matrix}{contrast matrix}
}
\value{
data.table
}
\description{
Group comparison for TMT data
}
\keyword{internal}
