% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonTMT.R
\name{MSstatsModerateTTest}
\alias{MSstatsModerateTTest}
\title{Moderate T statistic for group comparison}
\usage{
MSstatsModerateTTest(summarized, fitted_models, moderated)
}
\arguments{
\item{summarized}{protein-level data produced by the proteinSummarization function}

\item{fitted_models}{output of the MSstatsFitComparisonModelsTMT function}

\item{moderated}{if TRUE, moderation will be performed}
}
\value{
list
}
\description{
Moderate T statistic for group comparison
}
\keyword{internal}
