% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_normalize.R
\name{MSstatsNormalizeTMT}
\alias{MSstatsNormalizeTMT}
\title{Normalization for TMT data}
\usage{
MSstatsNormalizeTMT(input, type, normalize)
}
\arguments{
\item{input}{data.table}

\item{type}{"peptides" for peptide normalization between channel and run,
"proteins" for protein normalization}

\item{normalize}{logical, if TRUE, data will be normalized}
}
\value{
data.table
}
\description{
Normalization for TMT data
}
\keyword{internal}
