% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonTMT.R
\name{MSstatsPrepareForGroupComparisonTMT}
\alias{MSstatsPrepareForGroupComparisonTMT}
\title{Prepare output of proteinSummarization for group comparison}
\usage{
MSstatsPrepareForGroupComparisonTMT(
  input,
  remove_norm_channel,
  remove_empty_channel,
  call = rlang::caller_env()
)
}
\arguments{
\item{input}{output of proteinSummarization}

\item{remove_norm_channel}{if TRUE, "Norm" channel will be removed}

\item{remove_empty_channel, }{if TRUE, empty channel will be removed}
}
\value{
data.table
}
\description{
Prepare output of proteinSummarization for group comparison
}
\keyword{internal}
