% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparisonTMT.R
\name{MSstatsTestSingleProteinTMT}
\alias{MSstatsTestSingleProteinTMT}
\title{Hypothesis tests for a single protein in TMT data}
\usage{
MSstatsTestSingleProteinTMT(fitted_model, contrast_matrix)
}
\arguments{
\item{fitted_model}{single element of the MSstatsModerateTTest output}

\item{contrast_matrix}{contrast matrix}
}
\value{
list
}
\description{
Hypothesis tests for a single protein in TMT data
}
\keyword{internal}
