% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{OpenMStoMSstatsTMTFormat}
\alias{OpenMStoMSstatsTMTFormat}
\title{Generate MSstatsTMT required input format for OpenMS output}
\usage{
OpenMStoMSstatsTMTFormat(
  input,
  useUniquePeptide = TRUE,
  rmPSM_withfewMea_withinRun = TRUE,
  rmProtein_with1Feature = FALSE,
  summaryforMultiplePSMs = sum,
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL,
  ...
)
}
\arguments{
\item{input}{MSstatsTMT report from OpenMS}

\item{useUniquePeptide}{TRUE(default) removes peptides that are assigned for more than one proteins. We assume to use unique peptide for each protein.}

\item{rmPSM_withfewMea_withinRun}{TRUE (default) will remove the features that have 1 or 2 measurements within each Run.}

\item{rmProtein_with1Feature}{TRUE will remove the proteins which have only 1 peptide and charge. Defaut is FALSE.}

\item{summaryforMultiplePSMs}{sum(default) or max - when there are multiple measurements for certain feature in certain run, select the feature with the largest summation or maximal value.}

\item{use_log_file}{logical. If TRUE, information about data processing will be saved to a file.}

\item{append}{logical. If TRUE, information about data processing will be added to an existing log file.}

\item{verbose}{logical. If TRUE, information about data processing wil be printed to the console.}

\item{log_file_path}{character. Path to a file to which information about data processing will be saved. If not provided, such a file will be created automatically. If \code{append = TRUE}, has to be a valid path to a file.}

\item{...}{additional parameters to \code{data.table::fread}.}
}
\value{
\code{data.frame} of class \code{MSstatsTMT}.
}
\description{
Generate MSstatsTMT required input format for OpenMS output
}
\examples{
head(raw.om)
input.om <- OpenMStoMSstatsTMTFormat(raw.om)
head(input.om)

}
