% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{quant.msstats}
\alias{quant.msstats}
\title{Example of output from protein.summarizaiton function}
\format{A data frame with 100 rows and 7 variables.}
\usage{
quant.msstats
}
\description{
It is calculated from input.pd.
It is the output of protein.summarization function
It should includes the required columns as below.
The variables are as follows:
}
\details{
\itemize{
  \item Run : MS run ID
  \item Protein : Protein ID
  \item Abundance: Protein Abundance
  \item Channel : Labeling information (126, ... 131)
  \item Condition : Condition (ex. Healthy, Cancer, Time0)
  \item BioReplicate : Unique ID for biological subject.
  \item Mixture : Unique ID for TMT mixture.
}
}
\examples{
head(quant.msstats)

}
\keyword{datasets}
