% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MSstatsTMT.R
\docType{data}
\name{raw.pd}
\alias{raw.pd}
\title{Example of output from Proteome Discoverer 2.2 for TMT-10plex experiments.}
\format{
A data frame with 2858 rows and 50 variables.
}
\usage{
raw.pd
}
\description{
Example of Proteome discover PSM sheet.
It is the input for PDtoMSstatsTMTFormat function, with annotation file.
Annotation file should be made by users.
It includes peak intensities for 10 proteins
among 15 MS runs with TMT-10plex.
The variables are as follows:
}
\details{
\itemize{
\item Master.Protein.Accessions
\item Protein.Accessions
\item Annotated.Sequence
\item Charge
\item Ions.Score
\item Spectrum.File
\item Quan.Info
\item Channels : 126, ..., 131
}
}
\examples{
head(raw.pd)

}
\keyword{datasets}
