% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{required.input}
\alias{required.input}
\title{Example of output from PDtoMSstatsTMTFormat function}
\format{A data frame with 14740 rows and 8 variables.}
\usage{
required.input
}
\description{
It is formated from raw.input and annotation.
It is the output of PDtoMSstatsTMTFormat function and the input for protein.summarization function.
It should includes the required columns as below.
The variables are as follows:
}
\details{
\itemize{
  \item Protein : Protein ID
  \item PeptideSequence : Peptide sequence
  \item Charge : Ion charge
  \item PSM : combination of Peptide sequence and charge
  \item Channel : Labeling information (126, ... 131)
  \item Condition : Condition (ex. Healthy, Cancer, Time0)
  \item BioReplicate : Unique ID for biological subject.
  \item Run : MS run ID. It should be the same as Spectrum.File info in raw.input.
  \item Mixture : Mixture ID
  \item Intensity : Intensity
}
}
\examples{
head(required.input)

}
\keyword{datasets}
