% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MSstatsTMT.R
\docType{data}
\name{test.pairwise}
\alias{test.pairwise}
\title{Example of output from groupComparisonTMT function}
\format{
A data frame with 60 rows and 7 variables.
}
\usage{
test.pairwise
}
\description{
It is the output of groupComparisonTMT function,
which is made from \code{\link{quant.pd.msstats}}.
It is a list that consists of the following elements:
(1) ComparisonResult: statistical testing results;
(2) FittedModel: the fitted linear models
ComparisonResult should include the columns as below.
}
\details{
\itemize{
  \item Protein : Protein ID
  \item Label: Label of the pairwise comparision or contrast
  \item log2FC: Log2 fold change
  \item SE: Standard error of the comparsion of contrast results
  \item DF: Degree of freedom
  \item pvalue: Value of p statistic of the test
  \item adj.pvalue: adjusted p value
  \item issue: used for indicating the reason why a comparison is not testable. NA means the comparison is testable.
  'oneConditionMissing' means the protein has no measurements in one conndition of the comparison.
  Furtherone, when 'issue = oneConditionMissing', 'log2FC = Inf' means the negative condition
  (with coefficient -1 in the Label column)  is missing and 'log2FC = -Inf' means
  the positive condition (with coefficient 1 in the Label column)  is missing.
  completeMissing' means the protein has no measurements in all the connditions of the comparison.
  unfittableModel' means there is no enough measurements to fit the linear model.
  In other words, each condition has only one measurement.
}
}
\examples{
head(test.pairwise$ComparisonResult)

}
\keyword{internal}
