% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{annotation.mine}
\alias{annotation.mine}
\title{Example of annotation file for raw.mine, which is the output of SpectroMine.}
\format{A data frame with 72 rows and 5 variables.}
\usage{
annotation.mine
}
\description{
Annotation of example data, raw.mine, in this package.
It should be prepared by users.
The variables are as follows:
}
\details{
\itemize{
  \item Run : MS run ID. It should be the same as R.FileName info
  in raw.mine
  \item Channel : Labeling information (TMT6_126, ..., TMT6_131).
  The channels should be consistent with the channel columns in raw.mine.
  \item Condition : Condition (ex. Healthy, Cancer, Time0)
  \item Mixture : Mixture of samples with different TMT reagents, which can be analyzed in
  a single mass spectrometry experiment.
  The mixture may be fractionated into multiple fractions to increase the analytical depth.
  Then one mixture should correspond to multuple MS runs (One fraction is one MS run).
  For example, if 'Run' = 1, 2, 3 are three fractions of one TMT mixture of biological samples,
  then they should have same 'Mixture' value. Another special case is technical replicate.
  For example, if 'Run' = 4, 5, 6 are three technical replicates of one mixture, then they should match
  with same 'Mixture' value.
  \item BioReplicate : Unique ID for biological subject
}
}
\examples{
head(annotation.mine)

}
\keyword{datasets}
