% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{annotation.mq}
\alias{annotation.mq}
\title{Example of annotation file for raw.mq, which is the output of MaxQuant.}
\format{A data frame with 150 rows and 5 variables.}
\usage{
annotation.mq
}
\description{
Annotation of example data, raw.mq, in this package.
It should be prepared by users.
The variables are as follows:
}
\details{
\itemize{
  \item Run : MS run ID. It should be the same as Raw.file info
  in raw.mq
  \item Channel : Labeling information (channel.0, ..., channel.9).
  The channel index should be consistent with the channel columns in raw.mq.
  \item Condition : Condition (ex. Healthy, Cancer, Time0)
  \item Mixture : TMT Mixture ID. It is used to indicate the technical
  replicates or fractions. For example, if 'Run' 1, 2, 3 are the fractions
  of same TMT mixture, then they should have same 'Mixture'.
  \item BioReplicate : Unique ID for biological subject
}
}
\examples{
head(annotation.mq)

}
\keyword{datasets}
