% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{annotation.pd}
\alias{annotation.pd}
\title{Example of annotation file for raw.pd,
which is the PSM output of Proteome Discoverer}
\format{A data frame with 150 rows and 5 variables.}
\usage{
annotation.pd
}
\description{
Annotation of example data, raw.pd, in this package.
It should be prepared by users.
The variables are as follows:
}
\details{
\itemize{
  \item Run : MS run ID. It should be the same as Spectrum.File info
  in raw.pd.
  \item Channel : Labeling information (126, ... 131). It should be
  consistent with the channel columns in raw.pd.
  \item Condition : Condition (ex. Healthy, Cancer, Time0)
  \item Mixture : TMT Mixture ID. It is used to indicate the technical
  replicates or fractions. For example, if 'Run' 1, 2, 3 are the fractions
  of same TMT mixture, then they should have same 'Mixture'.
  \item BioReplicate : Unique ID for biological subject
}
}
\examples{
head(annotation.pd)

}
\keyword{datasets}
