% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{annotation.pd}
\alias{annotation.pd}
\title{Example of annotation file for raw.pd,
which is the PSM output of Proteome Discoverer}
\format{A data frame with 150 rows and 5 variables.}
\usage{
annotation.pd
}
\description{
Annotation of example data, raw.pd, in this package.
It should be prepared by users.
The variables are as follows:
}
\details{
\itemize{
  \item Run : MS run ID. It should be the same as Spectrum.File info
  in raw.pd.
  \item Channel : Labeling information (126, ... 131). It should be
  consistent with the channel columns in raw.pd.
  \item Condition : Condition (ex. Healthy, Cancer, Time0)
  \item Mixture : Mixture of samples with different TMT reagents, which can be analyzed in
  a single mass spectrometry experiment.
  The mixture may be fractionated into multiple fractions to increase the analytical depth.
  Then one mixture should correspond to multuple MS runs (One fraction is one MS run).
  For example, if 'Run' = 1, 2, 3 are three fractions of one TMT mixture of biological samples,
  then they should have same 'Mixture' value. Another special case is technical replicate.
  For example, if 'Run' = 4, 5, 6 are three technical replicates of one mixture, then they should match
  with same 'Mixture' value.
  \item BioReplicate : Unique ID for biological subject
}
}
\examples{
head(annotation.pd)

}
\keyword{datasets}
