% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{annotation}
\alias{annotation}
\title{Example of annotation file for raw.input}
\format{A data frame with 30 rows and 5 variables.}
\usage{
annotation
}
\description{
Annotation of example data, raw.input, in this package. It should be prepared by users.
The variables are as follows:
}
\details{
\itemize{
  \item Run : MS run ID
  \item Channel : Labeling ID
  \item Group : Condition (ex. Healthy, Cancer, Time0)
  \item BiologicalMixture : Mixture ID
  \item Subject : Unique ID for biological subject
}
}
\examples{
head(annotation)

}
\keyword{datasets}
