% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{.documentFunction}
\alias{.documentFunction}
\title{A dummy function to store shared documentation items.}
\usage{
.documentFunction(
  fewMeasurements,
  useUniquePeptide,
  summaryforMultipleRows,
  removeProtein_with1Feature,
  removeProtein_with1Protein,
  removeOxidationMpeptides,
  removeMpeptides
)
}
\arguments{
\item{fewMeasurements}{'remove'(default) will remove the features that have 1 or 2 measurements across runs.}

\item{useUniquePeptide}{TRUE (default) removes peptides that are assigned for more than one proteins. We assume to use unique peptide for each protein.}

\item{summaryforMultipleRows}{max(default) or sum - when there are multiple measurements for certain feature and certain run, use highest or sum of multiple intensities.}

\item{removeProtein_with1Feature}{TRUE will remove the proteins which have only 1 feature, which is the combination of peptide, precursor charge, fragment and charge. FALSE is default.}

\item{removeOxidationMpeptides}{TRUE will remove the peptides including 'oxidation (M)' in modification. FALSE is default.}

\item{removeMpeptides}{TRUE will remove the peptides including 'M' sequence. FALSE is default.}

\item{removeProtein_with1Peptide}{TRUE will remove the proteins which have only 1 peptide and charge. FALSE is default.}

\item{use_log_file}{logical. If TRUE, information about data processing will be saved to a file.}

\item{append}{logical. If TRUE, information about data processing will be added to an existing log file.}

\item{verbose}{logical. If TRUE, information about data processing wil be printed to the console.}

\item{log_file_path}{character. Path to a file to which information about data processing will be saved. If not provided, such a file will be created automatically. If \code{append = TRUE}, has to be a valid path to a file.}
}
\value{
NULL.
}
\description{
A dummy function to store shared documentation items.
}
\keyword{internal}
