% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{.getPhilosopherInput}
\alias{.getPhilosopherInput}
\title{Convert Philosopher parameters to consistent format}
\usage{
.getPhilosopherInput(input, path, folder)
}
\arguments{
\item{input}{list of tables exported by Philosopher. Fragpipe produces a csv file for each TMT mixture.}

\item{path}{a path to the folder with all the Philosopher msstats csv files. Fragpipe produces a msstats.csv file for each TMT mixture.}

\item{folder}{logical, if TRUE, path parameter will be treated as folder path and all msstats*.csv files will be imported. 
If FALSE, path parameter will be treated as a vector of fixed file paths.}
}
\description{
Convert Philosopher parameters to consistent format
}
\keyword{internal}
