% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{.getPhilosopherInput}
\alias{.getPhilosopherInput}
\title{Convert Philosopher parameters to consistent format}
\usage{
.getPhilosopherInput(input, path, folder)
}
\arguments{
\item{input}{data.frame of `msstats.csv` file produced by Philosopher}
}
\description{
Convert Philosopher parameters to consistent format
}
\keyword{internal}
