% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_checks.R
\name{.logSummarizationParams}
\alias{.logSummarizationParams}
\title{Log parameters for proteinSummarization function}
\usage{
.logSummarizationParams(
  method,
  global_norm,
  reference_norm,
  remove_norm_channel,
  remove_empty_channel
)
}
\arguments{
\item{method}{Four different summarization methods to protein-level
can be performed : "msstats"(default), "MedianPolish", "Median", "LogSum".}

\item{global_norm}{Global median normalization on peptide level data
(equalizing the medians across all the channels and MS runs).
Default is TRUE. It will be performed before protein-level summarization.}

\item{reference_norm}{Reference channel based normalization between MS runs
on protein level data.
\itemize{
\item TRUE (default): needs at least one reference channel in each MS run,
annotated by 'Norm' in Condition column. It will be performed after
protein-level summarization.
\item FALSE: no reference channel normalization.
If data only has one run, then reference_norm=FALSE.
}}

\item{remove_norm_channel}{TRUE (default) removes 'Norm' channels from
protein-level data.}

\item{remove_empty_channel}{TRUE (default) removes 'Empty'
channels from protein level data.}
}
\value{
TRUE invisibly after logging successfully
}
\description{
Log parameters for proteinSummarization function
}
\keyword{internal}
