% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_methods.R
\name{.medianPolish}
\alias{.medianPolish}
\title{Tukey median polish}
\usage{
.medianPolish(intensities, num_channels)
}
\arguments{
\item{intensities}{vector of log-intensities per protein and run}

\item{num_channels}{number of channels}
}
\value{
numeric vector with length `num_channels`
}
\description{
Tukey median polish
}
\keyword{internal}
