% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_normalize.R
\name{.normalizePeptides}
\alias{.normalizePeptides}
\title{Normalization between channels (before summarization)}
\usage{
.normalizePeptides(input, normalize)
}
\arguments{
\item{input}{data.table}

\item{normalize}{logical, if TRUE, `input` data will be normalized}
}
\value{
data.table
}
\description{
Normalization between channels (before summarization)
}
\keyword{internal}
