% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_checks.R
\name{.prepareForSummarization}
\alias{.prepareForSummarization}
\title{Prepare TMT data for protein-level summarization}
\usage{
.prepareForSummarization(input)
}
\arguments{
\item{input}{data.table}
}
\value{
data.table with required column types
}
\description{
Prepare TMT data for protein-level summarization
}
\keyword{internal}
