% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_checks.R
\name{.removeRedundantChannels}
\alias{.removeRedundantChannels}
\title{Remove empty and normalization channels}
\usage{
.removeRedundantChannels(input, remove_empty_channel, remove_norm_channel)
}
\arguments{
\item{input}{data.table processed by the protein summarization function}

\item{remove_empty_channel}{TRUE(default) removes 'Empty' channels from protein level data.}

\item{remove_norm_channel}{TRUE(default) removes 'Norm' channels from protein level data.}
}
\value{
data.table
}
\description{
Remove empty and normalization channels
}
\keyword{internal}
