% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_summarization_methods.R
\name{.summarizeTMT}
\alias{.summarizeTMT}
\title{Performs summarization for TMT data}
\usage{
.summarizeTMT(input, method, annotation)
}
\arguments{
\item{input}{data.table}

\item{method}{"mstats"/"MedianPolish"/"LogSum"/"Median"}

\item{annotation}{data.table with run and channel annotation}
}
\value{
data.table
}
\description{
Performs summarization for TMT data
}
\keyword{internal}
