% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{evidence}
\alias{evidence}
\title{Example of output from MaxQuant for TMT10 experiments.}
\format{A data frame with 1075 rows and 105 variables.}
\usage{
evidence
}
\description{
Example of MaxQuant peptide evidence file.
It is the input for MaxQtoMSstatsTMTFormat function, with proteinGroups.txt
 and annotation file.
It includes peak intensities for 10 proteins among 15 MS runs with TMT10.
The important variables are as follows:
}
\details{
\itemize{
  \item Proteins
  \item Protein.group.IDs
  \item Modified.sequence
  \item Charge
  \item Raw.file
  \item Score
  \item Potential.contaminant
  \item Reverse
  \item Channels : Reporter.intensity.corrected.0, ...,
  Reporter.intensity.corrected.9
}
}
\examples{
head(evidence)

}
\keyword{datasets}
