% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparsion.TMT.R
\name{groupComparison.TMT}
\alias{groupComparison.TMT}
\title{Group comparison on protein-level data}
\usage{
groupComparison.TMT(data, model = "proposed")
}
\arguments{
\item{data}{data: protein level data, which has columns Protein, Group, Subject, Run, Channel, IonIntensity}

\item{model}{Possible options: "proposed", "limma", "t"}
}
\value{
The sum of \code{x} and \code{y}.//TODO
}
\description{
Group comparison on protein-level data
}
