% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparsion.TMT.R
\name{groupComparison.TMT}
\alias{groupComparison.TMT}
\title{Finding differentially abundant proteins across conditions in TMT experiment}
\usage{
groupComparison.TMT(data, contrast.matrix = "pairwise",
  remove_norm_channel = TRUE, moderated = TRUE, adj.method = "BH")
}
\arguments{
\item{data}{Name of the output of protein.summarization function. It should have columns named Protein, Mixture, Run, Channel, Condition, BioReplicate, Abundance.}

\item{contrast.matrix}{Comparison between conditions of interests. 1) default is 'pairwise', which compare all possible pairs between two conditions. 2) Otherwise, users can specify the comparisons of interest. Based on the levels of conditions, specify 1 or -1 to the conditions of interests and 0 otherwise. The levels of conditions are sorted alphabetically.}

\item{remove_norm_channel}{TRUE(default) removes 'Norm' channels for inference step.}

\item{moderated}{Only for model = 'proposed'. If moderated = TRUE, then moderated t statistic will be calculated; otherwise, ordinary t statistic will be used.}

\item{adj.method}{adjusted method for multiple comparison. "BH" is default.}

\item{model}{Three different statistical approached can be performed : "proposed", "limma", "ttest". "proposed" is the default.}
}
\value{
data.frame with result of inference
}
\description{
Tests for significant changes in protein abundance across conditions based on a family of linear mixed-effects models in TMT experiment.
Experimental design of case-control study (patients are not repeatedly measured) or time course study (patients are repeatedly measured) is automatically determined based on proper statistical model.
}
\examples{
quant.byprotein <- protein.summarization(required.input,
                                         method="msstats",
                                         normalization=TRUE,)

test.byproposed <- groupComparison.TMT(quant.byprotein)

# Only compare condition 0.125 and 1
levels(quant.byprotein$Condition)
# 'Norm' should be not considered in the contrast
comparison<-matrix(c(-1,0,0,1),nrow=1)
# Set the names of each row
row.names(comparison)<-"1-0.125"
# Set the column names
colnames(comparison)<- c("0.125", "0.5", "0.667", "1")
test.contrast <- groupComparison.TMT(data = quant.byprotein, contrast.matrix = comparison)
}
