% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupComparsionTMT.R
\name{groupComparisonTMT}
\alias{groupComparisonTMT}
\title{Finding differentially abundant proteins across conditions in TMT experiment}
\usage{
groupComparisonTMT(data, contrast.matrix = "pairwise",
  moderated = FALSE, adj.method = "BH")
}
\arguments{
\item{data}{Name of the output of \code{\link{proteinSummarization}} function. It should have columns named Protein, Mixture, TechRepMixture, Run, Channel, Condition, BioReplicate, Abundance.}

\item{contrast.matrix}{Comparison between conditions of interests. 1) default is 'pairwise', which compare all possible pairs between two conditions. 2) Otherwise, users can specify the comparisons of interest. Based on the levels of conditions, specify 1 or -1 to the conditions of interests and 0 otherwise. The levels of conditions are sorted alphabetically.}

\item{moderated}{TRUE will moderate t statistic; FALSE (default) uses ordinary t statistic.}

\item{adj.method}{adjusted method for multiple comparison. "BH" is default.}
}
\value{
data.frame with result of inference
}
\description{
Tests for significant changes in protein abundance across conditions based on a family of linear mixed-effects models in TMT experiment.
Experimental design of case-control study (patients are not repeatedly measured) or time course study (patients are repeatedly measured) is automatically determined based on proper statistical model.
}
\examples{
data(input.pd)
# use protein.summarization() to get protein abundance data
quant.pd.msstats <- proteinSummarization(input.pd,
                                       method="msstats",
                                       normalization=TRUE)

test.pairwise <- groupComparisonTMT(quant.pd.msstats, moderated = TRUE)

# Only compare condition 0.125 and 1
levels(quant.pd.msstats$Condition)

# Compare condition 1 and 0.125
comparison<-matrix(c(-1,0,0,1),nrow=1)

# Set the names of each row
row.names(comparison)<-"1-0.125"

# Set the column names
colnames(comparison)<- c("0.125", "0.5", "0.667", "1")
test.contrast <- groupComparisonTMT(data = quant.pd.msstats,
contrast.matrix = comparison,
moderated = TRUE)
}
