% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{input.pd}
\alias{input.pd}
\title{Example of output from PDtoMSstatsTMTFormat function}
\format{A data frame with 20110 rows and 10 variables.}
\usage{
input.pd
}
\description{
It is calculated from raw.pd and annotation.pd
It is the output of PDtoMSstatsTMTFormat function
It should includes the required columns as below.
The variables are as follows:
}
\details{
\itemize{
  \item ProteinName : Protein ID
  \item PeptideSequence : peptide sequence
  \item Charge : peptide charge
  \item PSM : peptide ion and spectra match
  \item Channel : Labeling information (126, ... 131)
  \item Condition : Condition (ex. Healthy, Cancer, Time0)
  \item BioReplicate : Unique ID for biological subject.
  \item Run : MS run ID
  \item Mixture : Unique ID for TMT mixture.
  \item Intensity: Protein Abundance
}
}
\examples{
head(input.pd)

}
\keyword{datasets}
