% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinSummarization.R
\name{MSstatsSummarizationOutputTMT}
\alias{MSstatsSummarizationOutputTMT}
\title{Combine feature-level and protein-level data into single output}
\usage{
MSstatsSummarizationOutputTMT(
  summarized,
  processed,
  remove_empty_channel,
  remove_norm_channel
)
}
\arguments{
\item{summarized}{output of the getSummarizedTMT function}

\item{processed}{output of the getProcessedTMT function}

\item{remove_empty_channel}{TRUE(default) removes 'Empty' channels from protein level data.}

\item{remove_norm_channel}{TRUE(default) removes 'Norm' channels from protein level data.}
}
\value{
list that consists of two data.frames with feature-level and protein-level data
}
\description{
Combine feature-level and protein-level data into single output
}
\keyword{internal}
