% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSstatsTMT.R
\docType{package}
\name{MSstatsTMT}
\alias{MSstatsTMT}
\alias{MSstatsTMT-package}
\title{MSstatsTMT: A package for protein significance analysis in shotgun mass spectrometry-based proteomic experiments with tandem mass tag (TMT) labeling}
\description{
A set of tools for detecting differentially abundant peptides and proteins in shotgun mass spectrometry-based proteomic experiments with tandem mass tag (TMT) labeling.
}
\section{functions }{

\itemize{
  \item \code{\link{PDtoMSstatsTMTFormat}} : generating MSstatsTMT required input format for Proteome discoverer output.
  \item \code{\link{MaxQtoMSstatsTMTFormat}} : generating MSstatsTMT required input format for MaxQuant output.
  \item \code{\link{SpectroMinetoMSstatsTMTFormat}} : generating MSstatsTMT required input format for SpectroMine output.
  \item \code{\link{protein.summarization}} : summarizing PSM level quantification to protein level quantification.
  \item \code{\link{dataProcessPlots.TMT}} : visualizing for explanatory data analysis.
  \item \code{\link{groupComparison.TMT}} : testing for significant changes in protein abundance across conditions.
}
}

