% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDtoMSstatsTMTFormat.R
\name{PDtoMSstatsTMTFormat}
\alias{PDtoMSstatsTMTFormat}
\title{Converter for Proteome discoverer output}
\usage{
PDtoMSstatsTMTFormat(input, annotation, useNumProteinsColumn = TRUE,
  useUniquePeptide = TRUE, summaryforMultipleRows = max,
  removePSM_withMissingValue_withinRun = TRUE,
  removeProtein_with1Peptide = FALSE,
  which.proteinid = "Protein.Accessions")
}
\arguments{
\item{input}{output from Proteome discoverer : PSM sheet}
}
\value{
annotation data frame which contains column Run, Channel, Group, BiologicalRepliate. BiologicalRepliate indicates whether there is technical replicates or fractions.
}
\description{
Converter for Proteome discoverer output
}
