% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.summarization.R
\name{protein.summarization}
\alias{protein.summarization}
\title{Summarize PSM level data to protein level}
\usage{
protein.summarization(data, method = "MedianPolish")
}
\arguments{
\item{data}{PSM level data, which has columns Protein, PSM, Subject, Run, Channel, log2Intensity}

\item{method}{summarization methods. Possible options: "LogSum", "Median", "Biweight", "MedianPolish", "Huber"}
}
\value{
Protein Abundance
}
\description{
Summarize PSM level data to protein level
}
\examples{
head(required.input)
str(required.input)
MedianPolish.abun <- protein.summarization(required.input, "MedianPolish")
head(MedianPolish.abun)
}
