% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein.summarization.R
\name{protein.summarization}
\alias{protein.summarization}
\title{Summarize PSM level data to protein level}
\usage{
protein.summarization(data, annotation, method)
}
\arguments{
\item{data}{PSM level data, which has columns Protein, PSM, Subject, Run, Channel, IonIntensity}

\item{method}{summarization methods. Possible options: "LogSum", "Median", "Biweight", "MedianPolish", "Huber"}
}
\value{
The sum of \code{x} and \code{y}.//TODO
}
\description{
Summarize PSM level data to protein level
}
\examples{
TODO
}
