% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{proteinGroups}
\alias{proteinGroups}
\title{Example of proteinGroups file from MaxQuant for TMT-10plex experiments.}
\format{
A data frame with 1075 rows and 105 variables.
}
\usage{
proteinGroups
}
\description{
Example of proteinGroup.txt file from MaxQuant,
which is identified protein group information file.
It is the input for MaxQtoMSstatsTMTFormat function, with evidence.txt
and annotation file.
It includes identified protein groups for 10 proteins
among 15 MS runs with TMT10.
The important variables are as follows:
}
\details{
\itemize{
  \item id
  \item Protein.IDs
  \item Only.identified.by.site
  \item Potential.contaminant
  \item Reverse
}
}
\examples{
head(proteinGroups)

}
\keyword{datasets}
