% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_MSstatsTMT.R
\docType{data}
\name{quant.pd.msstats}
\alias{quant.pd.msstats}
\title{Example of output from proteinSummarizaiton function}
\format{
A data frame with 100 rows and 8 variables.
}
\usage{
quant.pd.msstats
}
\description{
It is made from \code{\link[=input.pd]{input.pd()}}.
It is the output of proteinSummarization function.
It is a list that consists of two data.frames with
feature-level (FeatureLevelData) and protein-level data (ProteinLevelData).
ProteinLevelData should include the required columns as below.
}
\details{
\itemize{
\item Run : MS run ID
\item Protein : Protein ID
\item Abundance: Protein-level summarized abundance
\item Channel : Labeling information (126, ... 131)
\item Condition : Condition (ex. Healthy, Cancer, Time0)
\item BioReplicate : Unique ID for biological subject.
\item TechRepMixture : Unique ID for technical replicate of one TMT mixture.
\item Mixture : Unique ID for TMT mixture.
}
}
\examples{
head(quant.pd.msstats$ProteinLevelData)

}
\keyword{internal}
