% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{raw.input}
\alias{raw.input}
\title{Raw input}
\format{A data frame with 1928 rows and 51 variables}
\usage{
raw.input
}
\description{
A dataset containing the raw input
The variables are as follows:
}
\details{
\itemize{
  \item Master.Protein.Accessions
  \item Protein.Accessions
}
}
\keyword{datasets}
