% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{raw.mine}
\alias{raw.mine}
\title{Example of output from SpectroMine for TMT-6plex experiments.}
\format{
A data frame with 170 rows and 28 variables.
}
\usage{
raw.mine
}
\description{
Example of SpectroMine PSM sheet.
It is the output of SpectroMine and the input for SpectroMinetoMSstatsTMTFormat function,
with annotation file.
Annotation file should be made by users.
It includes peak intensities for 10 proteins among 12 MS runs with TMT-6plex.
The important variables are as follows:
}
\details{
\itemize{
  \item PG.ProteinAccessions
  \item P.MoleculeID
  \item PP.Charge
  \item R.FileName
  \item PG.QValue
  \item PSM.Qvalue
  \item Channels : PSM.TMT6_126..Raw., ..., PSM.TMT6_131..Raw.
}
}
\examples{
head(raw.mine)

}
\keyword{datasets}
