% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{raw.om}
\alias{raw.om}
\title{Example of MSstatsTMT report from OpenMS for TMT-10plex experiments.}
\format{
A data frame with 860 rows and 13 variables.
}
\usage{
raw.om
}
\description{
Example of MSstatsTMT PSM sheet from MaxQuant.
It is the input for OpenMStoMSstatsTMTFormat function.
It includes peak intensities for 10 proteins among 27 MS runs from three TMT10 mixtures.
The important variables are as follows:
}
\details{
\itemize{
  \item RetentionTime
  \item ProteinName
  \item PeptideSequence
  \item Charge
  \item Channel
  \item Condition
  \item BioReplicate
  \item Run
  \item Mixture
  \item TechRepMixture
  \item Fraction
  \item Intensity
  \item Reference
}
}
\examples{
head(raw.om)

}
\keyword{datasets}
