% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectroMinetoMSstatsFormat.R
\name{SpectroMinetoMSstatsTMTFormat}
\alias{SpectroMinetoMSstatsTMTFormat}
\title{Generate MSstatsTMT required input format for SpectroMine output}
\usage{
SpectroMinetoMSstatsTMTFormat(input, annotation, fraction = FALSE,
  filter_with_Qvalue = TRUE, qvalue_cutoff = 0.01,
  useUniquePeptide = TRUE, rmPSM_withMissing_withinRun = FALSE,
  rmPSM_withfewMea_withinRun = TRUE, rmProtein_with1Feature = FALSE,
  summaryforMultipleRows = sum)
}
\arguments{
\item{input}{data name of SpectroMine PSM output. Read PSM sheet.}

\item{annotation}{data frame which contains column Run, Channel, Condition, BioReplicate, Mixture.}

\item{fraction}{indicates whether the data has fractions. If there are fractions, then overlapped peptide ions will be removed and then fractions are combined for each mixture.}

\item{filter_with_Qvalue}{TRUE(default) will filter out the intensities that have greater than qvalue_cutoff in EG.Qvalue column. Those intensities will be replaced with NA and will be considered as censored missing values for imputation purpose.}

\item{qvalue_cutoff}{Cutoff for EG.Qvalue. default is 0.01.}

\item{useUniquePeptide}{TRUE(default) removes peptides that are assigned for more than one proteins. We assume to use unique peptide for each protein.}

\item{rmPSM_withMissing_withinRun}{TRUE will remove PSM with any missing value within each Run. Defaut is FALSE.}

\item{rmPSM_withfewMea_withinRun}{only for rmPSM_withMissing_withinRun = FALSE. TRUE(default) will remove the features that have 1 or 2 measurements within each Run.}

\item{rmProtein_with1Feature}{TRUE will remove the proteins which have only 1 peptide and charge. Defaut is FALSE.}

\item{summaryforMultipleRows}{sum(default) or max - when there are multiple measurements for certain feature in certain run, select the feature with the largest summation or maximal value.}
}
\value{
input for protein.summarization function
}
\description{
Convert SpectroMine output into the required input format for MSstatsTMT.
}
\examples{
\dontrun{
head(raw.mine)
head(annotation.mine)
input.mine <- SpectroMinetoMSstatsTMTFormat(raw.mine, annotation.mine)
head(input.mine)
}
}
