% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{test.pairwise}
\alias{test.pairwise}
\title{Example of output from groupComparison.TMT function}
\format{A data frame with 60 rows and 7 variables.}
\usage{
test.pairwise
}
\description{
It is calculated from the result of protein.summarizaiton function with contrast.matrix = 'pairwise'.
It is the output of groupComparison.TMT function
It should includes the required columns as below.
The variables are as follows:
}
\details{
\itemize{
  \item Protein : Protein ID
  \item Label: Label of the pairwise comparision or contrast
  \item log2FC: Log2 fold change
  \item SE: Standard error of the comparsion of contrast results
  \item DF: Degree of freedom
  \item pvalue: Value of p statistic of the test
  \item adj.pvalue: adjusted p value
}
}
\examples{
head(test.pairwise)

}
\keyword{datasets}
